+++ src/main/java/org/icgc_argo/workflowgraphmanager/graphql/GraphLogDataFetcher.java
@@ -28,10 +28,10 @@ import java.util.List;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
-import org.icgc_argo.workflowgraphmanager.model.AggregationResult;
-import org.icgc_argo.workflowgraphmanager.model.GraphLog;
-import org.icgc_argo.workflowgraphmanager.model.SearchResult;
-import org.icgc_argo.workflowgraphmanager.model.Sort;
+import org.icgc_argo.workflowgraphmanager.graphql.model.AggregationResult;
+import org.icgc_argo.workflowgraphmanager.graphql.model.GraphLog;
+import org.icgc_argo.workflowgraphmanager.graphql.model.SearchResult;
+import org.icgc_argo.workflowgraphmanager.graphql.model.Sort;
 import org.icgc_argo.workflowgraphmanager.service.GraphLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
