+++ src/main/java/org/icgc_argo/workflowgraphmanager/graphql/model/Queue.java
@@ -16,45 +16,37 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc_argo.workflowgraphmanager.model;
+package org.icgc_argo.workflowgraphmanager.graphql.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import java.util.List;
 import java.util.Map;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NonNull;
 import lombok.SneakyThrows;
-import org.icgc_argo.workflowgraphmanager.model.base.Message;
+import org.icgc_argo.workflowgraphmanager.graphql.model.base.Message;
 import org.icgc_argo.workflowgraphmanager.utils.JacksonUtils;
 
 @Data
-@Builder
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
-public class Node {
-  @NonNull private String id;
+public class Queue {
+  private String id;
 
-  private Map<String, Object> config;
+  private String exchange;
 
-  private Boolean enabled;
+  private Pipeline pipeline;
 
-  private Integer capacity;
-
-  //  private String workflow; // todo: Make Workflow an entity
-
-  private String pipeline;
-
-  private List<Queue> queues;
+  private Node node;
 
   private List<Message> messages;
 
   private List<GraphLog> logs;
 
   @SneakyThrows
-  public static Node parse(@NonNull Map<String, Object> sourceMap) {
-    return JacksonUtils.parse(sourceMap, Node.class);
+  public static Queue parse(@NonNull Map<String, Object> sourceMap) {
+    return JacksonUtils.parse(sourceMap, Queue.class);
   }
 }
