+++ src/main/java/org/icgc_argo/workflowgraphmanager/graphql/model/SearchResult.java
@@ -16,7 +16,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.icgc_argo.workflowgraphmanager.model;
+package org.icgc_argo.workflowgraphmanager.graphql.model;
 
 import java.util.List;
 import lombok.Value;
