+++ src/main/java/org/icgc_argo/workflowgraphmanager/repository/GraphLogRepository.java
@@ -46,7 +46,7 @@ import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.elasticsearch.search.sort.FieldSortBuilder;
 import org.elasticsearch.search.sort.SortBuilders;
 import org.icgc_argo.workflowgraphmanager.config.ElasticsearchProperties;
-import org.icgc_argo.workflowgraphmanager.model.Sort;
+import org.icgc_argo.workflowgraphmanager.graphql.model.Sort;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
