+++ src/main/java/org/icgc_argo/workflowgraphmanager/service/GraphLogService.java
@@ -30,10 +30,10 @@ import java.util.Optional;
 import lombok.val;
 import org.elasticsearch.action.search.MultiSearchResponse;
 import org.elasticsearch.search.SearchHit;
-import org.icgc_argo.workflowgraphmanager.model.AggregationResult;
-import org.icgc_argo.workflowgraphmanager.model.GraphLog;
-import org.icgc_argo.workflowgraphmanager.model.SearchResult;
-import org.icgc_argo.workflowgraphmanager.model.Sort;
+import org.icgc_argo.workflowgraphmanager.graphql.model.AggregationResult;
+import org.icgc_argo.workflowgraphmanager.graphql.model.GraphLog;
+import org.icgc_argo.workflowgraphmanager.graphql.model.SearchResult;
+import org.icgc_argo.workflowgraphmanager.graphql.model.Sort;
 import org.icgc_argo.workflowgraphmanager.repository.GraphLogRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
